fix: storagev2 разбить класс по файлам
